Moved paper import out of dashboard and into seperate components. Added props to Card component to set background colour. Removed unused checkout component
