+++ src/components/Card.js
@@ -2,9 +2,10 @@ import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import CardContent from '@material-ui/core/CardContent';
 import Typography from '@material-ui/core/Typography';
+import Paper from '@material-ui/core/Paper';
 
-export default function SimpleCard({ color, number, text }) {
-	const useStyles = makeStyles({
+export default function SimpleCard({ background, color, number, text }) {
+	const useStyles = makeStyles((theme) => ({
 		root: {
 			width: '100%',
 		},
@@ -12,36 +13,45 @@ export default function SimpleCard({ color, number, text }) {
 			fontSize: 14,
 			fontWeight: 400,
 			marginTop: 10,
+			color: 'white',
 		},
 		number: {
 			fontSize: 60,
 			fontWeight: 700,
-			color: color,
+			color: 'white',
 		},
 		pos: {
 			marginBottom: 12,
 		},
-	});
+		paper: {
+			padding: theme.spacing(2),
+			textAlign: 'center',
+			color: 'white',
+			background: background,
+		},
+	}));
 	const classes = useStyles();
 
 	return (
 		<div className={classes.root}>
-			<CardContent>
-				<Typography
-					className={classes.number}
-					color="textSecondary"
-					gutterBottom
-				>
-					{number}
-				</Typography>
-				<Typography
-					className={classes.title}
-					color="textSecondary"
-					gutterBottom
-				>
-					{text}
-				</Typography>
-			</CardContent>
+			<Paper className={classes.paper}>
+				<CardContent>
+					<Typography
+						className={classes.number}
+						color="textSecondary"
+						gutterBottom
+					>
+						{number}
+					</Typography>
+					<Typography
+						className={classes.title}
+						color="textSecondary"
+						gutterBottom
+					>
+						{text}
+					</Typography>
+				</CardContent>
+			</Paper>
 		</div>
 	);
 }
