+++ src/components/UserDashboard.js
@@ -3,7 +3,6 @@ import { makeStyles } from '@material-ui/core/styles';
 import { useGlobalState } from '../config/globalState';
 import Bookings from './Bookings';
 import Continent from './Continent';
-import Paper from '@material-ui/core/Paper';
 import Grid from '@material-ui/core/Grid';
 import Card from './Card';
 import { getTeeth, getFairyDollars, getWishes } from '../utils/calculations';
@@ -21,11 +20,6 @@ const useStyles = makeStyles((theme) => ({
 	},
 	root: {
 		flexGrow: 1,
-	},
-	paper: {
-		padding: theme.spacing(2),
-		textAlign: 'center',
-		color: theme.palette.text.secondary,
 	},
 }));
 
@@ -50,67 +44,57 @@ function UserDashboard() {
 				<Grid container spacing={3}>
 					<Grid item xs={12} md={6} lg={3}>
 						{adminUser && (
-							<Paper className={classes.paper}>
-								<Card
-									number={`F$ ${getFairyDollars()}`}
-									text={'Current Fairy dollar per A$'}
-									color={'#ffc000'}
-								/>
-							</Paper>
+							<Card
+								number={`F$ ${getFairyDollars()}`}
+								text={'Current Fairy dollar per A$'}
+								background={
+									'radial-gradient(circle, rgba(107,236,252,1) 50%, rgba(56,205,212,1) 91%)'
+								}
+							/>
 						)}
 					</Grid>
 					<Grid item xs={12} md={6} lg={3}>
 						{adminUser && (
-							<Paper className={classes.paper}>
-								<Card
-									number={`F$ ${getFairyDollars()}`}
-									text={'Current Fairy dollar per A$'}
-									color={'black'}
-								/>
-							</Paper>
+							<Card
+								number={`F$ ${getFairyDollars()}`}
+								text={'Current Fairy dollar per A$'}
+								background={
+									'linear-gradient(180deg, rgba(255,205,241,1) 50%, rgba(235,173,237,1) 100%)'
+								}
+							/>
 						)}
 					</Grid>
 					<Grid item xs={12} md={6} lg={3}>
 						{adminUser && (
-							<Paper className={classes.paper}>
-								<Card
-									number={getTeeth(bookings)}
-									text={'Tooth exchanges today'}
-									color={'#ff3675'}
-								/>
-							</Paper>
+							<Card
+								number={getTeeth(bookings)}
+								text={'Tooth exchanges today'}
+								background={
+									'linear-gradient(94deg, rgba(81,27,119,1) 50%, rgba(41,20,115,1) 100%)'
+								}
+							/>
 						)}
 					</Grid>
 					<Grid item xs={12} md={6} lg={3}>
 						{adminUser && (
-							<Paper className={classes.paper}>
-								<Card
-									number={wishes}
-									text={'Wishes made in last minute'}
-									color={'#30d6af'}
-								/>
-							</Paper>
+							<Card
+								number={wishes}
+								text={'Wishes made in last minute'}
+								background={
+									'radial-gradient(circle, rgba(41,223,189,1) 62%, rgba(101,255,213,1) 96%)'
+								}
+							/>
 						)}
 					</Grid>
 					<Grid item xs={12} lg={6}>
-						{adminUser && (
-							<Paper className={classes.paper}>
-								<Continent />
-							</Paper>
-						)}
+						{adminUser && <Continent />}
 					</Grid>
 					<Grid item xs={12} lg={6}>
-						{adminUser && (
-							<Paper className={classes.paper}>
-								<Continent />
-							</Paper>
-						)}
+						{adminUser && <Continent />}
 					</Grid>
 
 					<Grid item xs={12}>
-						<Paper className={classes.paper}>
-							<Bookings bookings={bookings} />
-						</Paper>
+						<Bookings bookings={bookings} />
 					</Grid>
 				</Grid>
 			</div>
